Move the attachment logic into a separate file
